discarding scoring

+++ cms/grading/tasktypes/Communication.py
@@ -40,7 +40,7 @@ from cms.grading.steps import compilation_step,  evaluation_step_before_run, \
     evaluation_step_after_run, extract_outcome_and_text, \
     human_evaluation_message, merge_execution_stats, trusted_step
 from cms.grading.languagemanager import LANGUAGES, get_language
-from cms.grading.ParameterTypes import ParameterTypeInt
+from cms.grading.ParameterTypes import ParameterTypeChoice, ParameterTypeInt
 from cms.db import Executable
 from cms.grading.tasktypes import check_files_number
 from . import TaskType, check_executables_number, check_manager_present, \
@@ -59,15 +59,15 @@ class Communication(TaskType):
     """Task type class for tasks with a fully admin-controlled process.
 
     The task type will run *manager*, an admin-provided executable, and one or
-    more instances of the user solution, compiled together with a
+    more instances of the user solution, optionally compiled together with a
     language-specific stub.
 
     During the evaluation, the manager and each of the user processes
     communicate via FIFOs. The manager will read the input, send it (possibly
-    with some modifications) to the user process(es). The user processes, via
-    functions provided by the stub, will communicate with the manager. Finally,
-    the manager will decide outcome and text, and print them on stdout and
-    stderr.
+    with some modifications) to the user process(es). The user processes, either
+    via functions provided by the stub or by themselves, will communicate with
+    the manager. Finally, the manager will decide outcome and text, and print
+    them on stdout and stderr.
 
     The manager reads the input from stdin and writes to stdout and stderr the
     standard manager output (that is, the outcome on stdout and the text on
@@ -77,8 +77,10 @@ class Communication(TaskType):
     to a file named "output.txt"; the content of this file will be shown to
     users submitting a user test.
 
-    The stub receives as argument the fifos (from and to the manager) and if
-    there are more than one user processes, the 0-based index of the process.
+    The user process receives as argument the fifos (from and to the manager)
+    and, if there are more than one user processes, the 0-based index of the
+    process. The pipes can also be set up to be redirected to stdin/stdout: in
+    that case the names of the pipes are not passed as arguments.
 
     """
     # Filename of the manager (the stand-alone, admin-provided program).
@@ -93,6 +95,12 @@ class Communication(TaskType):
     # in case of a user test.
     OUTPUT_FILENAME = "output.txt"
 
+    # Constants used in the parameter definition.
+    COMPILATION_ALONE = "alone"
+    COMPILATION_STUB = "stub"
+    USER_IO_STD = "std_io"
+    USER_IO_FIFOS = "fifo_io"
+
     ALLOW_PARTIAL_SUBMISSION = False
 
     _NUM_PROCESSES = ParameterTypeInt(
@@ -100,7 +108,22 @@ class Communication(TaskType):
         "num_processes",
         "")
 
-    ACCEPTED_PARAMETERS = [_NUM_PROCESSES]
+    _COMPILATION = ParameterTypeChoice(
+        "Compilation",
+        "compilation",
+        "",
+        {COMPILATION_ALONE: "Submissions are self-sufficient",
+         COMPILATION_STUB: "Submissions are compiled with a stub"})
+
+    _USER_IO = ParameterTypeChoice(
+        "User I/O",
+        "user_io",
+        "",
+        {USER_IO_STD: "User processes read from stdin and write to stdout",
+         USER_IO_FIFOS: "User processes read from and write to fifos, "
+                        "whose paths are given as arguments"})
+
+    ACCEPTED_PARAMETERS = [_NUM_PROCESSES, _COMPILATION, _USER_IO]
 
     @property
     def name(self):
@@ -110,34 +133,42 @@ class Communication(TaskType):
     def __init__(self, parameters):
         super(Communication, self).__init__(parameters)
 
-        self.num_processes = 1
-        if len(self.parameters) > 0:
-            self.num_processes = self.parameters[0]
+        self.num_processes = self.parameters[0]
+        self.compilation = self.parameters[1]
+        self.io = self.parameters[2]
 
     def get_compilation_commands(self, submission_format):
         """See TaskType.get_compilation_commands."""
+        codenames_to_compile = []
+        if self._uses_stub():
+            codenames_to_compile.append(self.STUB_BASENAME + ".%l")
+        codenames_to_compile.extend(submission_format)
+        executable_filename = self._executable_filename(submission_format)
         res = dict()
         for language in LANGUAGES:
-            # Collect source filenames.
             source_ext = language.source_extension
-            source_filenames = [self.STUB_BASENAME + source_ext]
-            for codename in submission_format:
-                source_filenames.append(codename.replace(".%l", source_ext))
-            # Compute executable name.
-            executable_filename = self._executable_filename(submission_format)
-            # Build the compilation commands.
-            commands = language.get_compilation_commands(
-                source_filenames, executable_filename)
-            res[language.name] = commands
+            res[language.name] = language.get_compilation_commands(
+                [codename.replace(".%l", source_ext)
+                 for codename in codenames_to_compile],
+                executable_filename)
         return res
 
     def get_user_managers(self):
         """See TaskType.get_user_managers."""
-        return [self.STUB_BASENAME + ".%l"]
+        if self._uses_stub():
+            return [self.STUB_BASENAME + ".%l"]
+        else:
+            return []
 
     def get_auto_managers(self):
         """See TaskType.get_auto_managers."""
         return [self.MANAGER_FILENAME]
+
+    def _uses_stub(self):
+        return self.compilation == self.COMPILATION_STUB
+
+    def _uses_fifos(self):
+        return self.io == self.USER_IO_FIFOS
 
     @staticmethod
     def _executable_filename(codenames):
@@ -162,35 +193,37 @@ class Communication(TaskType):
 
         # Prepare the files to copy in the sandbox and to add to the
         # compilation command.
-        files_to_get = {}
-        source_filenames = []
+        filenames_to_compile = []
+        filenames_and_digests_to_get = {}
         # The stub, that must have been provided (copy and add to compilation).
-        stub_filename = self.STUB_BASENAME + source_ext
-        if not check_manager_present(job, stub_filename):
-            return
-        source_filenames.append(stub_filename)
-        files_to_get[stub_filename] = job.managers[stub_filename].digest
+        if self._uses_stub():
+            stub_filename = self.STUB_BASENAME + source_ext
+            if not check_manager_present(job, stub_filename):
+                return
+            filenames_to_compile.append(stub_filename)
+            filenames_and_digests_to_get[stub_filename] = \
+                job.managers[stub_filename].digest
         # User's submitted file(s) (copy and add to compilation).
         for codename, file_ in iteritems(job.files):
-            source_filename = codename.replace(".%l", source_ext)
-            source_filenames.append(source_filename)
-            files_to_get[source_filename] = file_.digest
+            filename = codename.replace(".%l", source_ext)
+            filenames_to_compile.append(filename)
+            filenames_and_digests_to_get[filename] = file_.digest
         # Any other useful manager (just copy).
         for filename, manager in iteritems(job.managers):
             if is_manager_for_compilation(filename, language):
-                files_to_get[filename] = manager.digest
+                filenames_and_digests_to_get[filename] = manager.digest
 
         # Prepare the compilation command
         executable_filename = self._executable_filename(iterkeys(job.files))
         commands = language.get_compilation_commands(
-            source_filenames, executable_filename)
+            filenames_to_compile, executable_filename)
 
         # Create the sandbox.
         sandbox = create_sandbox(file_cacher, name="compile")
         job.sandboxes.append(sandbox.get_root_path())
 
         # Copy all required files in the sandbox.
-        for filename, digest in iteritems(files_to_get):
+        for filename, digest in iteritems(filenames_and_digests_to_get):
             sandbox.create_file_from_storage(filename, digest)
 
         # Run the compilation.
@@ -210,7 +243,7 @@ class Communication(TaskType):
                 Executable(executable_filename, digest)
 
         # Cleanup.
-        delete_sandbox(sandbox, job.success)
+        delete_sandbox(sandbox, job.success, job.keep_sandbox)
 
     def evaluate(self, job, file_cacher):
         """See TaskType.evaluate."""
@@ -298,15 +331,23 @@ class Communication(TaskType):
 
         # Start the user submissions compiled with the stub.
         language = get_language(job.language)
+        main = self.STUB_BASENAME if self._uses_stub() else executable_filename
         processes = [None for i in indices]
         for i in indices:
-            args = [sandbox_fifo_manager_to_user[i],
-                    sandbox_fifo_user_to_manager[i]]
+            args = []
+            stdin_redirect = None
+            stdout_redirect = None
+            if self._uses_fifos():
+                args.extend([sandbox_fifo_manager_to_user[i],
+                             sandbox_fifo_user_to_manager[i]])
+            else:
+                stdin_redirect = sandbox_fifo_manager_to_user[i]
+                stdout_redirect = sandbox_fifo_user_to_manager[i]
             if self.num_processes != 1:
                 args.append(str(i))
             commands = language.get_evaluation_commands(
                 executable_filename,
-                main=self.STUB_BASENAME,
+                main=main,
                 args=args)
             # Assumes that the actual execution of the user solution is the
             # last command in commands, and that the previous are "setup"
@@ -319,6 +360,8 @@ class Communication(TaskType):
                 job.time_limit,
                 job.memory_limit,
                 dirs_map={fifo_dir[i]: (sandbox_fifo_dir[i], "rw")},
+                stdin_redirect=stdin_redirect,
+                stdout_redirect=stdout_redirect,
                 multiprocess=job.multithreaded_sandbox)
 
         # Wait for the processes to conclude, without blocking them on I/O.
@@ -351,7 +394,6 @@ class Communication(TaskType):
         # terminate correctly, we report an error (and no need for user stats).
         if not success:
             stats_user = None
-            pass
 
         # If just asked to execute, fill text and set dummy outcome.
         elif job.only_execution:
@@ -385,9 +427,9 @@ class Communication(TaskType):
         job.text = text
         job.plus = stats_user
 
-        delete_sandbox(sandbox_mgr, job.success)
+        delete_sandbox(sandbox_mgr, job.success, job.keep_sandbox)
         for s in sandbox_user:
-            delete_sandbox(s, job.success)
-        if not config.keep_sandbox:
+            delete_sandbox(s, job.success, job.keep_sandbox)
+        if job.success and not config.keep_sandbox and not job.keep_sandbox:
             for d in fifo_dir:
                 rmtree(d)
